% Generated by roxygen2 (4.0.1): do not edit by hand
\name{black.litterman}
\alias{black.litterman}
\title{Black Litterman Estimates}
\usage{
black.litterman(R, P, Mu, Sigma)
}
\arguments{
\item{R}{returns}

\item{P}{a K x N pick matrix}

\item{Mu}{vector of length N of the prior expected values. The sample mean
is used if \code{mu} is not provided as an argument.}

\item{Sigma}{an N x N matrix of the prior covariance matrix. The sample covariance
is used if \code{Sigma} is not provided as an argument.}
}
\value{
\itemize{
  \item{BLMu:}{ posterior expected values}
  \item{BLSigma:}{ posterior covariance matrix}
}
}
\description{
Compute the Black Litterman estimate of moments for the posterior normal.
}
\note{
This function is largely based on the work of Xavier Valls to port
the matlab code of Attilio Meucci to \R as documented in the Meucci package.
}
\author{
Ross Bennett, Xavier Valls
}
\references{
A. Meucci - "Exercises in Advanced Risk and Portfolio Management" \url{http://symmys.com/node/170}.
}
\seealso{
\code{\link{BlackLittermanFormula}}
}

