% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ac_ranking.R
\name{centroid.buckets}
\alias{centroid.buckets}
\title{Buckets Centroid}
\usage{
centroid.buckets(buckets, simulations = 1000)
}
\arguments{
\item{buckets}{a list where each element contains the index of the assets in
the respective bucket. The assets within each bucket have no order.
The bucket elements are in ascending order such that
R_bucket_1 < ... < R_bucket_n}

\item{simulations}{number of simulations}
}
\value{
the centroid vector
}
\description{
Compute the centroid for buckets of assets
}
\details{
A common use of buckets is to divide the assets into quartiles or deciles,
but is generalized here for an arbitrary number of buckets and arbitrary
number of assets in each bucket.
}
\author{
Ross Bennett
}

