\name{chart.RiskBudget}
\alias{chart.RiskBudget}
\title{Generic method to chart risk contribution}
\usage{
  chart.RiskBudget(object, ...)
}
\arguments{
  \item{object}{optimal portfolio object created by
  \code{\link{optimize.portfolio}}}

  \item{...}{passthrough parameters to \code{\link{plot}}}
}
\description{
  This function is the generic method to chart risk budget
  objectives for \code{optimize.portfolio} and
  \code{opt.list} objects.
}

