\name{chart.Scatter.DE}
\alias{chart.RiskReward}
\alias{chart.RiskReward.optimize.portfolio.DEoptim}
\alias{chart.RiskReward.optimize.portfolio.GenSA}
\alias{chart.RiskReward.optimize.portfolio.pso}
\alias{chart.RiskReward.optimize.portfolio.random}
\alias{chart.RiskReward.optimize.portfolio.ROI}
\alias{chart.Scatter.DE}
\alias{chart.Scatter.GenSA}
\alias{chart.Scatter.pso}
\alias{chart.Scatter.ROI}
\alias{chart.Scatter.RP}
\title{classic risk reward scatter}
\usage{
  chart.Scatter.DE(object, neighbors = NULL, ...,
    return.col = "mean", risk.col = "ES",
    chart.assets = FALSE, element.color = "darkgray",
    cex.axis = 0.8, xlim = NULL, ylim = NULL)

  chart.RiskReward.optimize.portfolio.DEoptim(object,
    neighbors = NULL, ..., return.col = "mean",
    risk.col = "ES", chart.assets = FALSE,
    element.color = "darkgray", cex.axis = 0.8,
    xlim = NULL, ylim = NULL)

  chart.Scatter.RP(object, neighbors = NULL, ...,
    return.col = "mean", risk.col = "ES",
    chart.assets = FALSE, element.color = "darkgray",
    cex.axis = 0.8, xlim = NULL, ylim = NULL)

  chart.RiskReward.optimize.portfolio.random(object,
    neighbors = NULL, ..., return.col = "mean",
    risk.col = "ES", chart.assets = FALSE,
    element.color = "darkgray", cex.axis = 0.8,
    xlim = NULL, ylim = NULL)

  chart.Scatter.ROI(object, neighbors = NULL, ...,
    rp = FALSE, return.col = "mean", risk.col = "ES",
    chart.assets = FALSE, element.color = "darkgray",
    cex.axis = 0.8, xlim = NULL, ylim = NULL)

  chart.RiskReward.optimize.portfolio.ROI(object,
    neighbors = NULL, ..., rp = FALSE, return.col = "mean",
    risk.col = "ES", chart.assets = FALSE,
    element.color = "darkgray", cex.axis = 0.8,
    xlim = NULL, ylim = NULL)

  chart.Scatter.pso(object, neighbors = NULL, ...,
    return.col = "mean", risk.col = "ES",
    chart.assets = FALSE, element.color = "darkgray",
    cex.axis = 0.8, xlim = NULL, ylim = NULL)

  chart.RiskReward.optimize.portfolio.pso(object,
    neighbors = NULL, ..., return.col = "mean",
    risk.col = "ES", chart.assets = FALSE,
    element.color = "darkgray", cex.axis = 0.8,
    xlim = NULL, ylim = NULL)

  chart.Scatter.GenSA(object, neighbors = NULL, ...,
    rp = FALSE, return.col = "mean", risk.col = "ES",
    chart.assets = FALSE, element.color = "darkgray",
    cex.axis = 0.8, ylim = NULL, xlim = NULL)

  chart.RiskReward.optimize.portfolio.GenSA(object,
    neighbors = NULL, ..., rp = FALSE, return.col = "mean",
    risk.col = "ES", chart.assets = FALSE,
    element.color = "darkgray", cex.axis = 0.8,
    ylim = NULL, xlim = NULL)

  chart.RiskReward(object, neighbors, ..., rp = FALSE,
    return.col = "mean", risk.col = "ES",
    element.color = "darkgray", cex.axis = 0.8,
    ylim = NULL, xlim = NULL)
}
\arguments{
  \item{object}{optimal portfolio created by
  \code{\link{optimize.portfolio}}}

  \item{neighbors}{set of 'neighbor' portfolios to
  overplot, see Details in \code{\link{charts.DE}}}

  \item{...}{any other passthru parameters}

  \item{rp}{TRUE/FALSE. If TRUE, random portfolios are
  generated by \code{\link{random_portfolios}} to view the
  feasible space}

  \item{return.col}{string matching the objective of a
  'return' objective, on vertical axis}

  \item{risk.col}{string matching the objective of a 'risk'
  objective, on horizontal axis}

  \item{chart.assets}{TRUE/FALSE. Includes a risk reward
  scatter of the assets in the chart}

  \item{cex.axis}{The magnification to be used for axis
  annotation relative to the current setting of \code{cex}}

  \item{element.color}{color for the default plot scatter
  points}

  \item{xlim}{set the x-axis limit, same as in
  \code{\link{plot}}}

  \item{ylim}{set the y-axis limit, same as in
  \code{\link{plot}}}
}
\description{
  classic risk reward scatter
}
\seealso{
  \code{\link{optimize.portfolio}}
}

