\name{chart.Weights.RP}
\alias{chart.Weights.RP}
\title{boxplot of the weight distributions in the random portfolios...}
\usage{chart.Weights.RP(RP, neighbors=NA, ...)}
\description{boxplot of the weight distributions in the random portfolios}
\seealso{\code{\link{random_portfolios}}}
\arguments{\item{RP}{set of random portfolios created by \code{\link{random_portfolios}}}
\item{neighbors}{set of 'neighbor portfolios to overplot}
\item{...}{any other passthru parameters}}
