\name{charts.RP}
\alias{charts.RP}
\title{scatter and weights chart  for random portfolios...}
\usage{charts.RP(RP, neighbors=NA, ...)}
\description{scatter and weights chart  for random portfolios}
\seealso{\code{\link{random_portfolios}}}
\arguments{\item{RP}{set of random portfolios created by \code{\link{random_portfolios}}}
\item{neighbors}{set of 'neighbor portfolios to overplot}
\item{...}{any other passthru parameters}}
