% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraint_v2}
\alias{constraint_v2}
\title{constructor for v2 constraint specification}
\usage{
constraint_v2(type, enabled = TRUE, ..., constrclass = "v2_constraint")
}
\arguments{
\item{type}{character type of the constraint to add or update}

\item{enabled}{TRUE/FALSE to enabled the constraint}

\item{constrclass}{name of class for the constraint}

\item{\dots}{any other passthru parameters}
}
\description{
See main documentation entry in \code{\link{add.constraint}}.
}

