% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRRA.R
\name{crra.RobustMoments}
\alias{crra.RobustMoments}
\title{# Custom moments fucntion for use in optimize portfolio
supplies moments to CRRA objective function use Boudt's statistical.factor.model function for fitting moments}
\usage{
crra.RobustMoments(R, k = 1, ...)
}
\arguments{
\item{R}{an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns}

\item{k}{number of factors to use, defaults to one}

\item{...}{passthrough to downastream methods}
}
\value{
list ovbject containing  mu, sigma, m3, m4
}
\description{
# Custom moments fucntion for use in optimize portfolio
supplies moments to CRRA objective function use Boudt's statistical.factor.model function for fitting moments
}
\examples{
data("edhec")
Moments <- crra.RobustMoments(edhec)
}
