% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMcd.moments.R
\name{custom.covMcd}
\alias{custom.covMcd}
\title{Compute moments using MCD robust method}
\usage{
custom.covMcd(R, ...)
}
\arguments{
\item{R}{xts object of asset returns}

\item{...}{parameters for covMcd}
}
\value{
a list contains mean and covariance matrix of the stock return matrix
}
\description{
This is a function that makes use of covMcd function from robustbase package. 
For more details of the parameters and function please refer to the manual of 
robustbase on cran.
}
