% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.covRob.R
\name{custom.covRob.Mcd}
\alias{custom.covRob.Mcd}
\title{Compute moments using covRob.MCD robust method}
\usage{
custom.covRob.Mcd(R, ...)
}
\arguments{
\item{R}{xts object of asset returns}

\item{...}{parameters for covRob.Mcd}
}
\value{
a list contains mean and covariance matrix of the stock return matrix
}
\description{
This is a function that makes use of covMcd function from robustbase package. 
For more details of the parameters and function please refer to the manual of 
robustbase on cran.
}
