% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.covRob.R
\name{custom.covRob.MM}
\alias{custom.covRob.MM}
\title{Compute moments using covRob.MM methods}
\usage{
custom.covRob.MM(R, ...)
}
\arguments{
\item{R}{xts object of asset returns}

\item{...}{parameters for covRob.MM}
}
\value{
a list contains mean and covariance matrix of the data matrix
}
\description{
This is a function that makes use of covRob from package RobStatTM to calculate
robust estimators of moments, including mean and variance.
}
\author{
Yifu Kang
}
