% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRob.moments.R
\name{custom.covRob}
\alias{custom.covRob}
\title{Compute moments using covRob methods}
\usage{
custom.covRob(R, ...)
}
\arguments{
\item{R}{xts object of asset returns}

\item{...}{parameters for covRob}
}
\value{
a list contains mean and covariance matrix of the data matrix
}
\description{
This is a function that makes use of covRob from package RobStatTM to calculate
robust estimators of moments, including mean and variance.
}
