% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.covRob.R
\name{custom.covRob.Rocke}
\alias{custom.covRob.Rocke}
\title{Compute returns mean vector and covariance matrix with custom.covRob.Rocke}
\usage{
custom.covRob.Rocke(R, ...)
}
\arguments{
\item{R}{xts object of asset returns}

\item{...}{parameters for covRob.Rocke}
}
\value{
a list containing covariance matrix sigma and mean vector mu
}
\description{
custom.covRob.Rocke uses the RobStatTM package function covRobRocke to compute a robust
mean vector and robust covariance matrix for a portfolio's asset returns
}
\details{
For parameter details, see covRobRocke in the RobStatTM Reference
Manual at \url{https://cran.r-project.org/web/packages/RobStatTM/index.html}
}
\author{
Yifu Kang
}
