% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRob.moments.R
\name{custom.covRobMM}
\alias{custom.covRobMM}
\title{Compute moments using covRobMM methods}
\usage{
custom.covRobMM(R, ...)
}
\arguments{
\item{R}{xts object of asset returns}

\item{...}{parameters for covRobMM}
}
\value{
a list contains mean and covariance matrix of the data matrix
}
\description{
This is a function that makes use of covRobMM from package RobStatTM to calculate
robust estimators of moments, including mean and variance.
}
\author{
Yifu Kang
}
