% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRob.moments.R
\name{custom.covRobRocke}
\alias{custom.covRobRocke}
\title{Compute moments using covRobRocke methods}
\usage{
custom.covRobRocke(R, ...)
}
\arguments{
\item{R}{xts object of asset returns}

\item{...}{parameters for covRobRocke}
}
\value{
a list contains mean and covariance matrix of the data matrix
}
\description{
This is a function that makes use of covRobRocke from package RobStatTM to calculate
robust estimators of moments, including mean and variance.
}
\author{
Yifu Kang
}
