% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsgs.moments.R
\name{custom.TSGS}
\alias{custom.TSGS}
\title{Compute TSGS moments}
\usage{
custom.TSGS(R, ...)
}
\arguments{
\item{R}{xts object of asset returns}

\item{...}{parameters for tsgs}
}
\value{
a list contains mean and covariance matrix of the stock return matrix
}
\description{
This is a function making use of TSGS function from GSE package to compute
the Two-Step Generalized S-Estimate, a robust estimate of location 
and scatter for data with cell-wise and case-wise contamination.
}
\references{
Claudio Agostinelli, Andy Leung, "Robust estimation of multivariate 
            location and scatter in the presence of cellwise and casewise contamination",
            2014.
}
