\name{diversification}
\alias{diversification}
\title{Function to compute diversification as a constraint}
\usage{
  diversification(weights)
}
\arguments{
  \item{weights}{vector of asset weights}
}
\description{
  Diversification is defined as 1 minus the sum of the
  squared weights \deqn{diversification = 1 - sum(w^2)}
}
\author{
  Ross Bennett
}

