% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_fn_map.R
\name{.is_projection_feasible}
\alias{.is_projection_feasible}
\title{Check if a weight vector is feasible w.r.t. all projection constraints}
\usage{
.is_projection_feasible(
  w,
  min_sum,
  max_sum,
  min_box,
  max_box,
  groups = NULL,
  cLO = NULL,
  cUP = NULL,
  ftol = 1e-08
)
}
\description{
Check if a weight vector is feasible w.r.t. all projection constraints
}
\keyword{internal}
