\name{extract.cokurtosis}
\alias{extract.cokurtosis}
\title{Cokurtosis Estimate}
\usage{
  extract.cokurtosis(model, ...)
}
\arguments{
  \item{model}{statistical factor model estimated via
  \code{\link{statistical.factor.model}}}

  \item{\dots}{not currently used}
}
\value{
  cokurtosis matrix estimate
}
\description{
  Extract the cokurtosis matrix estimate from a statistical
  factor model
}
\author{
  Ross Bennett
}
\seealso{
  \code{\link{statistical.factor.model}}
}

