\name{extract.coskewness}
\alias{extract.coskewness}
\title{Coskewness Estimate}
\usage{
  extract.coskewness(model, ...)
}
\arguments{
  \item{model}{statistical factor model estimated via
  \code{\link{statistical.factor.model}}}

  \item{\dots}{not currently used}
}
\value{
  coskewness matrix estimate
}
\description{
  Extract the coskewness matrix estimate from a statistical
  factor model
}
\author{
  Ross Bennett
}
\seealso{
  \code{\link{statistical.factor.model}}
}

