% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.factor.model.R
\name{extractCokurtosis}
\alias{extractCokurtosis}
\title{Cokurtosis Estimate}
\usage{
extractCokurtosis(model, ...)
}
\arguments{
\item{model}{statistical factor model estimated via 
\code{\link{statistical.factor.model}}}

\item{\dots}{not currently used}
}
\value{
cokurtosis matrix estimate
}
\description{
Extract the cokurtosis matrix estimate from a statistical factor model
}
\seealso{
\code{\link{statistical.factor.model}}
}
\author{
Ross Bennett
}
