% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.factor.model.R
\name{extractCovariance}
\alias{extractCovariance}
\title{Covariance Estimate}
\usage{
extractCovariance(model, ...)
}
\arguments{
\item{model}{statistical factor model estimated via 
\code{\link{statistical.factor.model}}}

\item{\dots}{not currently used}
}
\value{
covariance matrix estimate
}
\description{
Extract the covariance matrix estimate from a statistical factor model
}
\seealso{
\code{\link{statistical.factor.model}}
}
\author{
Ross Bennett
}
