% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractstats.R
\name{extractFeasibility}
\alias{extractFeasibility}
\alias{extractFeasibility.optimize.portfolio}
\alias{extractFeasibility.optimize.portfolio.rebalancing}
\alias{extractFeasibility.opt.list}
\alias{extractFeasibility.opt.rebal.list}
\title{Extract feasibility reports from optimization results}
\usage{
extractFeasibility(object, ...)

\method{extractFeasibility}{optimize.portfolio}(object, ..., as.data.frame = FALSE)

\method{extractFeasibility}{optimize.portfolio.rebalancing}(object, ..., as.data.frame = FALSE)

\method{extractFeasibility}{opt.list}(object, ..., as.data.frame = FALSE)

\method{extractFeasibility}{opt.rebal.list}(object, ..., as.data.frame = FALSE)
}
\arguments{
\item{object}{An optimization result object.}

\item{...}{Additional arguments passed to methods.}

\item{as.data.frame}{logical. If \code{TRUE}, the report(s) are converted
to tidy data frames via \code{\link{as.data.frame.feasibility_report}}.
Default \code{FALSE}.}
}
\value{
A \code{feasibility_report} object, a list of such objects,
  or a data frame when \code{as.data.frame = TRUE}.
}
\description{
Generic function to extract \code{feasibility_report} objects from
\code{optimize.portfolio} results, including rebalancing and
multi-portfolio objects.
}
\examples{
\dontrun{
opt <- optimize.portfolio(R, portf, optimize_method = "ROI")
extractFeasibility(opt)
extractFeasibility(opt, as.data.frame = TRUE)
}

}
\seealso{
\code{\link{check_portfolio_feasibility}},
  \code{\link{as.data.frame.feasibility_report}}
}
