% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractObjectiveMeasures}
\alias{extractObjectiveMeasures}
\title{Extract the objective measures}
\usage{
extractObjectiveMeasures(object)
}
\arguments{
\item{object}{list returned by optimize.portfolio}
}
\value{
list of objective measures
}
\description{
This function will extract the objective measures from the optimal portfolio
run via \code{optimize.portfolio}
}
\author{
Ross Bennett
}
\seealso{
\code{\link{optimize.portfolio}}
}

