\name{extractstats}
\alias{extractstats}
\title{extract some stats from a DEoptim portfolio list fun via foreach...}
\usage{extractstats(resultlist)}
\description{extract some stats from a DEoptim portfolio list fun via foreach}
\details{TODO: Rewrite this function to test the input object and direct to the correct parse function}
\seealso{\code{\link{optimize.portfolio}}}
\arguments{\item{resultlist}{list returned by optimize.portfolio}}
