\name{extractStats.rp}
\alias{extractStats.rp}
\title{unlist random portfolio resultsoops...}
\usage{extractStats.rp(OptimResults, ...)}
\description{unlist random portfolio resultsoops}
\details{This just flattens the $random_portfolio_objective_results part of the object}
\seealso{\code{\link{random_portfolios}}}
\arguments{\item{OptimResults}{list returned by optimize.portfolio}
\item{...}{any other passthru parameters}}
