\name{extractWeights.optimize.portfolio}
\alias{extractWeights.optimize.portfolio}
\title{extract weights from output of optimize.portfolio}
\usage{
  extractWeights.optimize.portfolio(object)
}
\arguments{
  \item{object}{object of type optimize.portfolio to
  extract weights from}
}
\description{
  extract weights from output of optimize.portfolio
}
\author{
  Ross Bennett
}
\seealso{
  \code{\link{optimize.portfolio}}
}

