\name{extractWeights.rebal}
\alias{extractWeights.rebal}
\title{extract time series of weights from output of \link{optimize...}
\usage{extractWeights.rebal(RebalResults, ...)}
\description{extract time series of weights from output of \code{\link{optimize.portfolio.rebalancing}}}
\details{TODO make this compatible with extractStats functions}
\seealso{\code{\link{optimize.portfolio.rebalancing}}}
\arguments{\item{RebalResults}{object of type optimize.portfolio.rebalancing to extract weights from}
\item{...}{any other passthru parameters}}
