\name{HHI}
\alias{HHI}
\title{Function to calculate weight concentration}
\usage{
  HHI(weights, groups = NULL)
}
\arguments{
  \item{weights}{}

  \item{groups}{}
}
\description{
  This function calculates the concentration of weights
  using the Herfindahl–Hirschman Index as a measure of
  concentration
}
\author{
  Ross Bennett
}

