\name{insert_constraints}
\alias{insert_constraints}
\title{Insert a list of constraints into the constraints slot of a portfolio object}
\usage{
  insert_constraints(portfolio, constraints)
}
\arguments{
  \item{portfolio}{object of class 'portfolio'}

  \item{constraints}{list of constraint objects}
}
\description{
  Insert a list of constraints into the constraints slot of
  a portfolio object
}
\author{
  Ross Bennett
}

