\name{insert_objectives}
\alias{insert_objectives}
\title{Insert a list of objectives into the objectives slot of a portfolio object}
\usage{
  insert_objectives(portfolio, objectives)
}
\arguments{
  \item{portfolio}{object of class 'portfolio'}

  \item{objectives}{list of objective objects}
}
\description{
  Insert a list of objectives into the objectives slot of a
  portfolio object
}
\author{
  Ross Bennett
}

