% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{is.constraint}
\alias{is.constraint}
\title{check function for constraints}
\usage{
is.constraint(x)
}
\arguments{
\item{x}{object to test for type \code{constraint}}
}
\description{
check function for constraints
}
\author{
Brian G. Peterson
}
