% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/objective.R
\name{is.objective}
\alias{is.objective}
\title{check class of an objective object}
\usage{
is.objective(x)
}
\arguments{
\item{x}{an object potentially of type 'objective' to test}
}
\description{
check class of an objective object
}
\author{
Brian G. Peterson
}

