% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_failure.R
\name{is.optimization_failure}
\alias{is.optimization_failure}
\title{Test if an object is an optimization failure}
\usage{
is.optimization_failure(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Logical; \code{TRUE} if \code{x} inherits from
  \code{"optimization_failure"}.
}
\description{
Test if an object is an optimization failure
}
\seealso{
\code{\link{optimization_failure}}
}
