% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/portfolio.R
\name{is.portfolio}
\alias{is.portfolio}
\title{check function for portfolio}
\usage{
is.portfolio(x)
}
\arguments{
\item{x}{object to test for type \code{portfolio}}
}
\description{
check function for portfolio
}
\author{
Ross Bennett
}

