% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optFUN.R
\name{maxret_opt}
\alias{maxret_opt}
\title{Maximum Return LP Optimization}
\usage{
maxret_opt(R, moments, constraints, target, solver = "glpk", control = NULL)
}
\arguments{
\item{R}{xts object of asset returns}

\item{moments}{object of moments computed based on objective functions}

\item{constraints}{object of constraints in the portfolio object extracted with \code{get_constraints}}

\item{target}{target return value}

\item{solver}{solver to use}

\item{control}{list of solver control parameters}
}
\description{
This function is called by optimize.portfolio to solve maximum return
}
\author{
Ross Bennett
}

