% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.efficient.frontier.R
\name{meaneqs.efficient.frontier}
\alias{meaneqs.efficient.frontier}
\title{Generate the efficient frontier for a mean-EQS portfolio xinran}
\usage{
meaneqs.efficient.frontier(portfolio, R, n.portfolios = 25, ...)
}
\arguments{
\item{portfolio}{}

\item{R}{}

\item{n.portfolios}{}

\item{...}{}
}
\value{

}
\description{
Generate the efficient frontier for a mean-EQS portfolio xinran
}
