% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extractstats.R
\name{name.replace}
\alias{name.replace}
\title{utility function to replace awkward named from unlist}
\usage{
name.replace(rnames)
}
\arguments{
\item{rnames}{character vector of names to check for cleanup}
}
\description{
utility function to replace awkward named from unlist
}

