% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRRA.R
\name{pasd}
\alias{pasd}
\title{Objective function for  to compute annualized standard deviation for MONTHLY data}
\usage{
pasd(R, weights, sigma, N = 36)
}
\arguments{
\item{R}{list of monthlu returns}

\item{weights}{character vector of weights}

\item{sigma}{cov(R)}

\item{N}{Number of months default 36}
}
\value{
matrix with objective value
}
\description{
Objective function for  to compute annualized standard deviation for MONTHLY data
}
\examples{
data("edhec")
pasd(R = edhec, weights = rep(1/ncol(edhec), times = 13 ), stats::cov(edhec), N = 36  )
}
