\name{plot.optimize.portfolio.DEoptim}
\alias{plot.optimize.portfolio}
\alias{plot.optimize.portfolio.DEoptim}
\alias{plot.optimize.portfolio.GenSA}
\alias{plot.optimize.portfolio.pso}
\alias{plot.optimize.portfolio.random}
\alias{plot.optimize.portfolio.ROI}
\title{plot method for optimize.portfolio.DEoptim output}
\usage{
  \method{plot}{optimize.portfolio.DEoptim} (x, ...,
    return.col = "mean", risk.col = "ES",
    chart.assets = FALSE, neighbors = NULL,
    main = "optimized portfolio plot", xlim = NULL,
    ylim = NULL)

  \method{plot}{optimize.portfolio.random} (x, ...,
    return.col = "mean", risk.col = "ES",
    chart.assets = FALSE, neighbors = NULL, xlim = NULL,
    ylim = NULL, main = "optimized portfolio plot")

  \method{plot}{optimize.portfolio} (x, ...,
    return.col = "mean", risk.col = "ES",
    chart.assets = FALSE, neighbors = NULL, xlim = NULL,
    ylim = NULL, main = "optimized portfolio plot")

  \method{plot}{optimize.portfolio.ROI} (x, ...,
    rp = FALSE, risk.col = "ES", return.col = "mean",
    chart.assets = FALSE, element.color = "darkgray",
    neighbors = NULL, main = "ROI.Portfolios", xlim = NULL,
    ylim = NULL)

  \method{plot}{optimize.portfolio.pso} (x, ...,
    return.col = "mean", risk.col = "ES",
    chart.assets = FALSE, cex.axis = 0.8,
    element.color = "darkgray", neighbors = NULL,
    main = "PSO.Portfolios", xlim = NULL, ylim = NULL)

  \method{plot}{optimize.portfolio.GenSA} (x, ...,
    rp = FALSE, return.col = "mean", risk.col = "ES",
    chart.assets = FALSE, cex.axis = 0.8,
    element.color = "darkgray", neighbors = NULL,
    main = "GenSA.Portfolios", xlim = NULL, ylim = NULL)
}
\arguments{
  \item{x}{set of portfolios created by
  \code{\link{optimize.portfolio}}}

  \item{...}{any other passthru parameters}

  \item{return.col}{string name of column to use for
  returns (vertical axis)}

  \item{risk.col}{string name of column to use for risk
  (horizontal axis)}

  \item{chart.assets}{TRUE/FALSE to include risk-return
  scatter of assets}

  \item{neighbors}{set of 'neighbor portfolios to overplot}

  \item{main}{an overall title for the plot: see
  \code{\link{title}}}

  \item{xlim}{set the limit on coordinates for the x-axis}

  \item{ylim}{set the limit on coordinates for the y-axis}

  \item{x}{set of portfolios created by
  \code{\link{optimize.portfolio}}}

  \item{...}{any other passthru parameters}

  \item{return.col}{string name of column to use for
  returns (vertical axis)}

  \item{risk.col}{string name of column to use for risk
  (horizontal axis)}

  \item{chart.assets}{TRUE/FALSE to include risk-return
  scatter of assets}

  \item{neighbors}{set of 'neighbor portfolios to overplot}

  \item{xlim}{set the limit on coordinates for the x-axis}

  \item{ylim}{set the limit on coordinates for the y-axis}

  \item{main}{an overall title for the plot: see
  \code{\link{title}}}

  \item{x}{set of portfolios created by
  \code{\link{optimize.portfolio}}}

  \item{...}{any other passthru parameters}

  \item{return.col}{string name of column to use for
  returns (vertical axis)}

  \item{risk.col}{string name of column to use for risk
  (horizontal axis)}

  \item{chart.assets}{TRUE/FALSE to include risk-return
  scatter of assets}

  \item{neighbors}{set of 'neighbor portfolios to overplot}

  \item{xlim}{set the limit on coordinates for the x-axis}

  \item{ylim}{set the limit on coordinates for the y-axis}

  \item{main}{an overall title for the plot: see
  \code{\link{title}}}

  \item{x}{object created by
  \code{\link{optimize.portfolio}}}

  \item{...}{any other passthru parameters}

  \item{rp}{TRUE/FALSE to plot feasible portfolios
  generated by \code{\link{random_portfolios}}}

  \item{risk.col}{string matching the objective of a 'risk'
  objective, on horizontal axis}

  \item{return.col}{string matching the objective of a
  'return' objective, on vertical axis}

  \item{chart.assets}{TRUE/FALSE to include risk-return
  scatter of assets}

  \item{element.color}{color for the default plot scatter
  points}

  \item{neighbors}{set of 'neighbor' portfolios to
  overplot}

  \item{main}{an overall title for the plot: see
  \code{\link{title}}}

  \item{xlim}{set the limit on coordinates for the x-axis}

  \item{ylim}{set the limit on coordinates for the y-axis}

  \item{x}{object created by
  \code{\link{optimize.portfolio}}}

  \item{...}{any other passthru parameters}

  \item{return.col}{string matching the objective of a
  'return' objective, on vertical axis}

  \item{risk.col}{string matching the objective of a 'risk'
  objective, on horizontal axis}

  \item{chart.assets}{TRUE/FALSE to include risk-return
  scatter of assets}

  \item{cex.axis}{The magnification to be used for axis
  annotation relative to the current setting of \code{cex}}

  \item{element.color}{color for the default plot scatter
  points}

  \item{neighbors}{set of 'neighbor' portfolios to
  overplot}

  \item{main}{an overall title for the plot: see
  \code{\link{title}}}

  \item{xlim}{set the limit on coordinates for the x-axis}

  \item{ylim}{set the limit on coordinates for the y-axis}

  \item{x}{object created by
  \code{\link{optimize.portfolio}}}

  \item{...}{any other passthru parameters}

  \item{rp}{TRUE/FALSE to plot feasible portfolios
  generated by \code{\link{random_portfolios}}}

  \item{return.col}{string matching the objective of a
  'return' objective, on vertical axis}

  \item{risk.col}{string matching the objective of a 'risk'
  objective, on horizontal axis}

  \item{chart.assets}{TRUE/FALSE to include risk-return
  scatter of assets}

  \item{cex.axis}{The magnification to be used for axis
  annotation relative to the current setting of \code{cex}}

  \item{element.color}{color for the default plot scatter
  points}

  \item{neighbors}{set of 'neighbor' portfolios to
  overplot}

  \item{main}{an overall title for the plot: see
  \code{\link{title}}}

  \item{xlim}{set the limit on coordinates for the x-axis}

  \item{ylim}{set the limit on coordinates for the y-axis}
}
\description{
  Scatter and weights chart for DEoptim portfolio
  optimizations run with trace=TRUE

  Scatter and weights chart for random portfolio
  optimizations run with trace=TRUE

  Scatter and weights chart for portfolio optimization

  Scatter and weights chart for ROI portfolio optimizations
  run with trace=TRUE

  Scatter and weights chart for pso portfolio optimizations
  run with trace=TRUE

  Scatter and weights chart for GenSA portfolio
  optimizations run with trace=TRUE
}
\details{
  \code{return.col} must be the name of a function used to
  compute the return metric on the random portfolio weights
  \code{risk.col} must be the name of a function used to
  compute the risk metric on the random portfolio weights

  \code{neighbors} may be specified in three ways. The
  first is as a single number of neighbors.  This will
  extract the \code{neighbors} closest portfolios in terms
  of the \code{out} numerical statistic. The second method
  consists of a numeric vector for \code{neighbors}. This
  will extract the \code{neighbors} with portfolio index
  numbers that correspond to the vector contents. The third
  method for specifying \code{neighbors} is to pass in a
  matrix. This matrix should look like the output of
  \code{\link{extractStats}}, and should contain
  \code{risk.col},\code{return.col}, and weights columns
  all properly named.

  \code{return.col} must be the name of a function used to
  compute the return metric on the random portfolio
  weights. \code{risk.col} must be the name of a function
  used to compute the risk metric on the random portfolio
  weights.

  \code{neighbors} may be specified in three ways. The
  first is as a single number of neighbors.  This will
  extract the \code{neighbors} closest portfolios in terms
  of the \code{out} numerical statistic. The second method
  consists of a numeric vector for \code{neighbors}. This
  will extract the \code{neighbors} with portfolio index
  numbers that correspond to the vector contents. The third
  method for specifying \code{neighbors} is to pass in a
  matrix. This matrix should look like the output of
  \code{\link{extractStats}}, and should contain
  \code{risk.col},\code{return.col}, and weights columns
  all properly named.

  This is a fallback that will be called for classes of
  portfolio that do not have specific pre-existing plot
  methods.

  \code{neighbors} may be specified in three ways. The
  first is as a single number of neighbors.  This will
  extract the \code{neighbors} closest portfolios in terms
  of the \code{out} numerical statistic. The second method
  consists of a numeric vector for \code{neighbors}. This
  will extract the \code{neighbors} with portfolio index
  numbers that correspond to the vector contents. The third
  method for specifying \code{neighbors} is to pass in a
  matrix. This matrix should look like the output of
  \code{\link{extractStats}}, and should contain
  \code{risk.col},\code{return.col}, and weights columns
  all properly named.

  The ROI optimizers do not store the portfolio weights
  like DEoptim or random portfolios random portfolios can
  be generated for the scatter plot.

  \code{return.col} must be the name of a function used to
  compute the return metric on the random portfolio
  weights. \code{risk.col} must be the name of a function
  used to compute the risk metric on the random portfolio
  weights

  \code{return.col} must be the name of a function used to
  compute the return metric on the random portfolio
  weights. \code{risk.col} must be the name of a function
  used to compute the risk metric on the random portfolio
  weights.

  \code{return.col} must be the name of a function used to
  compute the return metric on the random portfolio
  weights. \code{risk.col} must be the name of a function
  used to compute the risk metric on the random portfolio
  weights.
}
\author{
  Ross Bennett

  Ross Bennett

  Ross Bennett
}
\seealso{
  \code{\link{optimize.portfolio}}

  \code{\link{optimize.portfolio}}

  \code{\link{optimize.portfolio}}
}

