% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{print.constraint}
\alias{print.constraint}
\title{print method for constraint objects}
\usage{
\method{print}{constraint}(x, ...)
}
\arguments{
\item{x}{object of class \code{constraint}}

\item{\dots}{any other passthru parameters}
}
\description{
print method for constraint objects
}
\author{
Ross Bennett
}

