\name{print.optimize.portfolio.GenSA}
\alias{print.optimize.portfolio.GenSA}
\title{Printing Output of optimize.portfolio}
\usage{
  print.optimize.portfolio.GenSA(x, ...,
    digits = max(3, getOption("digits") - 3))
}
\arguments{
  \item{x}{an object of class
  \code{optimize.portfolio.GenSA} resulting from a call to
  \code{\link{optimize.portfolio}}}

  \item{digits}{the number of significant digits to use
  when printing}

  \item{...}{any other passthru parameters}
}
\description{
  print method for optimize.portfolio.GenSA
}

