\name{print.optimize.portfolio.ROI}
\alias{print.optimize.portfolio.ROI}
\title{Printing Output of optimize.portfolio}
\usage{
  \method{print}{optimize.portfolio.ROI} (x, ...,
    digits = max(3, getOption("digits") - 3))
}
\arguments{
  \item{x}{an object of class \code{optimize.portfolio.ROI}
  resulting from a call to
  \code{\link{optimize.portfolio}}}

  \item{...}{any other passthru parameters}

  \item{digits}{the number of significant digits to use
  when printing.}
}
\description{
  print method for optimize.portfolio.ROI
}
\author{
  Ross Bennett
}

