\name{print.summary.optimize.portfolio}
\alias{print.summary.optimize.portfolio}
\title{Summarizing Output of optimize.portfolio}
\usage{
  \method{print}{summary.optimize.portfolio} (x, ...)
}
\arguments{
  \item{object}{an object of class "optimize.portfolio.pso"
  resulting from a call to optimize.portfolio}

  \item{...}{any other passthru parameters. Currently not
  used.}
}
\description{
  summary method for class "optimize.portfolio"
}
\author{
  Ross Bennett
}

