\name{random_walk_portfolios}
\alias{random_walk_portfolios}
\title{deprecated random portfolios wrapper until we write a random trades function...}
\usage{random_walk_portfolios(...)}
\description{deprecated random portfolios wrapper until we write a random trades function}
\details{TODO: write a function for random trades that only makes n trades and increases/decreases other elements to compensate.
random_walk_portfolios(seed,y,min_sum=.99,max_sum=1.01,permutations=100,max_permutations=5000)}
\author{bpeterson}
\arguments{\item{...}{}}
