% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.moments.r
\name{ranking.moments}
\alias{ranking.moments}
\title{Moment ranking function for either}
\usage{
ranking.moments(R, n = 1, method = c("meucci", "ac"))
}
\arguments{
\item{R}{an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns}

\item{n}{numeric lookback period for ranking}

\item{method}{character either "meucci" or "ac"}
}
\value{
list ovbject containing  mu, sigma
}
\description{
Meucci: Fully Flexible Views Framework
Almgren and Chriss(AC): Portfolios from Sorts
}
\details{
The moment function assumes that the assets with the highest return will continue to outperform
}
\examples{
data("edhec")
AC_Moments <- ranking.moments(edhec,n=3, method  = "ac")
meucci_Moments <- ranking.moments(edhec,n=3, method  = "meucci")
}
