% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.portfolio.R
\name{reduce_matrix_mincov}
\alias{reduce_matrix_mincov}
\title{helper function to drop columns of matrix until all covariances have have least mincov data points in their estimation}
\usage{
reduce_matrix_mincov(mat, mincov = 52, verbose = TRUE)
}
\arguments{
\item{mat}{a matrix (usually xts) of returns with observations on rows and assets on columns}

\item{mincov}{the minimum number of observations that must be present in all variances and covariances of
a given asset for it to be retained in the returned matrix. Any asset having fewer than \code{mincov}
observations is dropped from the returned returns matrix.}

\item{verbose}{A TRUE/FALSE argument indicating whether to print to screen which assets were dropped}
}
\value{
A modified returns matrix with assets dropped if they have fewer than \code{mincov} observations for any
  variance or covariance cell
}
\description{
helper function to drop columns of matrix until all covariances have have least mincov data points in their estimation
}
\author{
Michael Hallquist
}
