\name{return_objective}
\alias{return_objective}
\title{constructor for class return_objective...}
\usage{return_objective(name, enabled=FALSE, ..., multiplier=-1, target)}
\description{constructor for class return_objective}
\author{bpeterson}
\arguments{\item{name}{name of the objective, should correspond to a function, though we will try to make allowances}
\item{enabled}{TRUE/FALSE}
\item{\dots}{any other passthru parameters}
\item{multiplier}{multiplier to apply to the objective, usually 1 or -1}
\item{target}{univariate target for the objective}}
