\name{summary.efficient.frontier}
\alias{summary.efficient.frontier}
\title{Summarize an efficient frontier object}
\usage{
  summary.efficient.frontier(object, ..., digits = 3)
}
\arguments{
  \item{x}{objective of class \code{efficient.frontier}}

  \item{...}{passthrough parameters}
}
\description{
  Summary method for efficient frontier objects. Display
  the call to create or extract the efficient frontier
  object as well as the weights and risk and return metrics
  along the efficient frontier.
}
\author{
  Ross Bennett
}

