% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.optimize.portfolio}
\alias{summary.optimize.portfolio}
\title{Summarizing output of optimize.portfolio}
\usage{
\method{summary}{optimize.portfolio}(object, ...)
}
\arguments{
\item{object}{an object of class \code{optimize.portfolio}.}

\item{...}{any other passthru parameters. Currently not used.}
}
\description{
summary method for class \code{optimize.portfolio}
}
\author{
Ross Bennett
}
\seealso{
\code{\link{optimize.portfolio}}
}

