\name{summary.optimize.portfolio.rebalancing}
\alias{summary.optimize.portfolio.rebalancing}
\title{summary method for optimize...}
\usage{summary.optimize.portfolio.rebalancing(object, ...)}
\description{summary method for optimize.portfolio.rebalancing}
\arguments{\item{object}{object of type optimize.portfolio.rebalancing}
\item{\dots}{any other passthru parameters}}
