% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customMomentTrackingError.r
\name{te.target}
\alias{te.target}
\title{Tracking Error Custom Objective Function}
\source{
\url{https://rossb34.github.io/PortfolioAnalyticsPresentation2017/#29}
}
\usage{
te.target(R, Rb, weights, min.te = 0.02, max.te = 0.05, scale = 12)
}
\arguments{
\item{R}{an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns}

\item{Rb}{an xts, vector, matrix, data frame, timeSeries or zoo object of benchmark returns}

\item{weights}{weights vector}

\item{min.te}{a numeric value for the minimum annualized Tracking Error}

\item{max.te}{a numeric value for the maximum annualized Tracking Error}

\item{scale}{a numeric value how many rows there are per year/period.}
}
\value{

}
\description{
Define a custom objective function for penalized tracking error
}
\examples{
#using the final column of R als benchmark for the sake of simplicity
data("edhec")
Equalweights <- rep(1/(ncol(edhec)-1), ncol(edhec)-1)
test <- te.target( R = edhec[,1:12], Rb = edhec[,13]*10,  weights = Equalweights
                 , min.te = 0.02, max.te = 0.05, scale = 12) 
}
