% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.portfolio.R
\name{trim_portfolio}
\alias{trim_portfolio}
\title{function to trim a portfolio of assets with missing data between start and end that would
invalidate optimization
helper function to trim a portfolio containing missing assets within the interval of interest}
\usage{
trim_portfolio(portfolio, returns, start = NULL, end = NULL, mincov = 10)
}
\arguments{
\item{portfolio}{portfolio.spec object to be modified}

\item{start}{character vector start date to be used to window \code{returns}}

\item{end}{character vector end date to be used to window \code{returns}}

\item{mincov}{number of observations that must be present in variances and covariances for an asset
to be retained in the resulting portfolio object}
}
\value{
time series of returns to be checked for missingness

A modified \code{portfolio.spec} object containing only assets with observations above the \code{mincov}
  threshold in covariance estimation.
}
\description{
function to trim a portfolio of assets with missing data between start and end that would
invalidate optimization
helper function to trim a portfolio containing missing assets within the interval of interest
}
\details{
Portfolio optimization breaks down if there is missingness in the assets being considered within the window of
  interest. Although the moments of returns are calculated using pairwise completeness, the variance-covariance matrix
  still breaks down if there is no overlap in two assets (e.g., if one product is introduced and another is phased out).
  
  To overcome this, the \code{reduce_matrix_mincov} function trims the returns matrix to assets that have at least \code{mincov}
  observations. However, if we need to regenerate a corresponding portfolio specification, we must trim out the bad assets
  from the set, then re-add the constraints and objectives. This function achieves that purpose.
  
  It accepts an existing portfolio specification object and a (possibly) reduced \code{returns} matrix. The \code{returns} matrix
  is checked for a minimum number of observations using \code{reduce_matrix_mincov}. If any assets are dropped from the returns,
  a new, respecified portfolio object is generated only containing the assets with sufficient data for covariance estimation.
  Constraints and objectives are then added from the original \code{portfolio} and the object is returned to the user.
}
\author{
Michael Hallquist
}
