% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsgs.moments.R
\name{tsgs.moments}
\alias{tsgs.moments}
\title{Compute TSGS moments}
\usage{
tsgs.moments(
  R,
  filter = "UBF-DDC",
  partial.impute = FALSE,
  tol = 1e-04,
  maxiter = 150,
  method = "bisquare",
  init = "emve_c"
)
}
\arguments{
\item{R}{xts object of asset returns}

\item{filter}{the filter to be used in the first step. Available choices are 
"UBF-DDC","UBF","DDC","UF". The default one is "UBF-DDC".}

\item{partial.impute}{whether partial imputation is used prior to estimation.
The default is FALSE.}

\item{tol}{tolerance for the convergence criterion. Default is 1e-4.}

\item{maxiter}{maximum number of iterations. Default is 150.}

\item{method}{loss function to use, "bisquare" or "rocke". Default is "bisquare"}

\item{init}{type of initial estimator. Options include "emve", "qc", "huber","imputed","emve_c"}
}
\value{
estimators of first and second moments
}
\description{
This is a function making use of TSGS function from GSE package to compute
the Two-Step Generalized S-Estimate, a robust estimate of location 
and scatter for data with cell-wise and case-wise contamination.
}
\references{
Claudio Agostinelli, Andy Leung, "Robust estimation of multivariate 
            location and scatter in the presence of cellwise and casewise contamination",
            2014.
}
