\name{update.constraint}
\alias{update.constraint}
\title{function for updating constrints, not well tested, may be broken...}
\usage{update.constraint(object, ...)}
\description{function for updating constrints, not well tested, may be broken}
\details{can we use the generic update.default function?}
\author{bpeterson}
\arguments{\item{object}{}
\item{...}{}}
