% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.portfolio.R
\name{zero_assets}
\alias{zero_assets}
\title{small helper function to zero out certain assets that cannot be in the optimization (due to missingness)}
\usage{
zero_assets(portfolio, to_drop)
}
\description{
small helper function to zero out certain assets that cannot be in the optimization (due to missingness)
}
