% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.plot.R
\name{backtest.plot}
\alias{backtest.plot}
\title{generate plots of the cumulative returns and drawdown for back-testing}
\usage{
backtest.plot(
  R,
  log_return = FALSE,
  plotType = "both",
  colorSet = NULL,
  ltySet = NULL,
  lwdSet = NULL
)
}
\arguments{
\item{R}{an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns}

\item{log_return}{arithmetic return or log return, the default is arithmetic return}

\item{plotType}{ret, drawdown, or the default is both}

\item{colorSet}{users can design the color by providing a vector of color}

\item{ltySet}{users can design lty by providing a vector of lty}

\item{lwdSet}{users can design lwd by providing a vector of lwd}
}
\description{
generate plots of the cumulative returns and drawdown for back-testing
}
\author{
Peter Carl, Xinran Zhao, Yifu Kang
}
