% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.plot.R
\name{backtest.plot}
\alias{backtest.plot}
\title{generate plots of the cumulative returns and drawdown for back-testing}
\usage{
backtest.plot(
  R,
  log_return = FALSE,
  field = "both",
  colorset = NULL,
  ltyset = NULL
)
}
\arguments{
\item{R}{an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns}

\item{log_return}{arithmetic return or log return, the default is arithmetic return}

\item{field}{ret, drawdown, or the default is both}

\item{colorset}{users can design the color by providing a vector of color}

\item{ltyset}{users can design lty by providing a vector of lty}
}
\description{
generate plots of the cumulative returns and drawdown for back-testing
}
\author{
Peter Carl, Xinran Zhao, Yifu Kang
}
