\name{chart.Scatter.ROI}
\alias{chart.Scatter.ROI}
\title{classic risk return scatter of random portfolios}
\usage{
  chart.Scatter.ROI(ROI, neighbors = NULL, ..., rp = FALSE,
    return.col = "mean", risk.col = "ES",
    element.color = "darkgray", cex.axis = 0.8)
}
\arguments{
  \item{ROI}{object created by
  \code{\link{optimize.portfolio}}}

  \item{rp}{matrix of random portfolios generated by
  \code{\link{random_portfolio}}}

  \item{return.col}{string matching the objective of a
  'return' objective, on vertical axis}

  \item{risk.col}{string matching the objective of a 'risk'
  objective, on horizontal axis}

  \item{...}{any other passthru parameters}

  \item{cex.axis}{The magnification to be used for axis
  annotation relative to the current setting of \code{cex}}

  \item{element.color}{color for the default plot scatter
  points}
}
\description{
  The ROI optimizers do not store the portfolio weights
  like DEoptim or random portfolios so we will generate
  random portfolios for the scatter plot.
}
\details{
  \code{return.col} must be the name of a function used to
  compute the return metric on the random portfolio weights
  \code{risk.col} must be the name of a function used to
  compute the risk metric on the random portfolio weights
}
\author{
  Ross Bennett
}
\seealso{
  \code{\link{optimize.portfolio}}
}

