\name{chart.Scatter.RP}
\alias{chart.Scatter.RP}
\title{classic risk return scatter of random portfolios...}
\usage{chart.Scatter.RP(RP, neighbors=NA, ...)}
\description{classic risk return scatter of random portfolios}
\seealso{\code{\link{random_portfolios}}}
\arguments{\item{RP}{set of random portfolios created by \code{\link{random_portfolios}}}
\item{neighbors}{set of 'neighbor portfolios to overplot}
\item{...}{any other passthru parameters}}
