\name{chart.Weights.EF}
\alias{chart.Weights.EF}
\title{Chart weights along an efficient frontier}
\usage{
  chart.Weights.EF(object, ...)
}
\arguments{
  \item{object}{object to chart}

  \item{\dots}{any other passthru parameters}
}
\description{
  This function is a generic method to chart weights along
  an efficient frontier
}

