\name{charts.RP}
\alias{charts.RP}
\title{scatter and weights chart  for random portfolios...}
\usage{charts.RP(RP, risk.col, return.col, neighbors=NA, las=3, main="Random Portfolios", ...)}
\description{scatter and weights chart  for random portfolios}
\seealso{\code{\link{optimize.portfolio}}}
\arguments{\item{RP}{set of random portfolios created by \code{\link{optimize.portfolio}}}
\item{neighbors}{set of 'neighbor portfolios to overplot}
\item{...}{any other passthru parameters}}
