% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.covRob.R
\name{custom.covRob.Mcd}
\alias{custom.covRob.Mcd}
\title{Compute returns mean vector and covariance matrix with custom.covRob.Mcd}
\usage{
custom.covRob.Mcd(R, ...)
}
\arguments{
\item{R}{xts object of asset returns}

\item{...}{parameters for covRob.Mcd}
}
\value{
a list containing covariance matrix sigma and mean vector mu
}
\description{
custom.covRob.Mcd uses the robustbase package function covMcd to compute a robust
mean vector and robust covariance matrix for a portfolio's asset returns
}
\details{
For parameter details, see covMcd in the robustbase Reference
Manual at \url{https://cran.r-project.org/web/packages/robustbase/index.html}
}
