% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.covRob.R
\name{custom.covRob.MM}
\alias{custom.covRob.MM}
\title{Compute returns mean vector and covariance matrix with covRob.MM}
\usage{
custom.covRob.MM(R, ...)
}
\arguments{
\item{R}{xts object of asset returns}

\item{...}{parameters for covRob.MM}
}
\value{
a list containing mean and a covariance matrix
}
\description{
covRob.MM uses the RobStatTM package function covRobMM to compute a robust
mean vector and robust covariance matrix for a portfolio's asset returns
}
\details{
For parameter details, see covRobMM in the RobStatTM Reference
Manual at \url{https://cran.r-project.org/web/packages/RobStatTM/index.html}
}
\author{
Yifu Kang
}
