% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractrisk.R
\name{extract_risk}
\alias{extract_risk}
\title{extract the risk value when knowing the weights}
\usage{
extract_risk(
  R,
  w,
  ES_alpha = 0.05,
  CSM_alpha = 0.05,
  EQS_alpha = 0.05,
  moment_setting = NULL
)
}
\arguments{
\item{R}{an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns}

\item{w}{the weight of the portfolio}

\item{ES_alpha}{the default value is 0.05, but could be specified as any value between 0 and 1}

\item{CSM_alpha}{the default value is 0.05, but could be specified as any value between 0 and 1}

\item{EQS_alpha}{the default value is 0.05, but could be specified as any value between 0 and 1}

\item{moment_setting}{the default is NULL, should provide moment_setting=list(mu=, sigma=) if customize momentFUN}
}
\description{
extract the risk value when knowing the weights
}
