% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractrisk.R
\name{extract_risk}
\alias{extract_risk}
\title{extract the risk value when knowing the weights}
\usage{
extract_risk(R, w, portfolio = NULL)
}
\arguments{
\item{R}{an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns}

\item{w}{the weight of the portfolio}

\item{portfolio}{an object of type "portfolio" specifying the constraints and objectives for the optimization}
}
\description{
extract the risk value when knowing the weights
}
