% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractCokurtosis}
\alias{extractCokurtosis}
\title{Cokurtosis Estimate}
\usage{
extractCokurtosis(model, ...)
}
\arguments{
\item{model}{statistical factor model estimated via
\code{\link{statistical.factor.model}}}

\item{\dots}{not currently used}
}
\value{
cokurtosis matrix estimate
}
\description{
Extract the cokurtosis matrix estimate from a statistical factor model
}
\author{
Ross Bennett
}
\seealso{
\code{\link{statistical.factor.model}}
}

