% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractCoskewness}
\alias{extractCoskewness}
\title{Coskewness Estimate}
\usage{
extractCoskewness(model, ...)
}
\arguments{
\item{model}{statistical factor model estimated via
\code{\link{statistical.factor.model}}}

\item{\dots}{not currently used}
}
\value{
coskewness matrix estimate
}
\description{
Extract the coskewness matrix estimate from a statistical factor model
}
\author{
Ross Bennett
}
\seealso{
\code{\link{statistical.factor.model}}
}

