% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stat.factor.model.R
\name{extractCoskewness}
\alias{extractCoskewness}
\title{Coskewness Estimate}
\usage{
extractCoskewness(model, ...)
}
\arguments{
\item{model}{statistical factor model estimated via
\code{\link{statistical.factor.model}}}

\item{\dots}{not currently used}
}
\value{
coskewness matrix estimate
}
\description{
Extract the coskewness matrix estimate from a statistical factor model
}
\author{
Ross Bennett
}
\seealso{
\code{\link{statistical.factor.model}}
}

