\name{extractStats.optimize.portfolio.random}
\alias{extractStats.optimize.portfolio.random}
\title{unlist random portfolio resultsoops...}
\usage{extractStats.optimize.portfolio.random(object, prefix, ...)}
\description{unlist random portfolio resultsoops}
\details{This just flattens the $random_portfolio_objective_results part of the object}
\seealso{\code{\link{optimize.portfolio}}
\code{\link{random_portfolios}}
\code{\link{extractStats}}}
\arguments{\item{object}{list returned by optimize.portfolio}
\item{prefix}{prefix to add to output row names}
\item{...}{any other passthru parameters}}
