\name{extractWeights.optimize.portfolio}
\alias{extractWeights.optimize.portfolio}
\title{extract weights from output of optimize.portfolio}
\usage{
  extractWeights.optimize.portfolio(object, ...)
}
\arguments{
  \item{object}{object of class \code{optimize.portfolio}
  to extract weights from}

  \item{...}{passthrough parameters. Not currently used}
}
\description{
  extract weights from output of optimize.portfolio
}
\author{
  Ross Bennett
}
\seealso{
  \code{\link{optimize.portfolio}}
}

