% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/objectiveFUN.R
\name{HHI}
\alias{HHI}
\title{Concentration of weights}
\usage{
HHI(weights, groups = NULL)
}
\arguments{
\item{weights}{set of portfolio weights}

\item{groups}{list of vectors of grouping}
}
\description{
This function computes the concentration of weights using the Herfindahl Hirschman Index
}
\author{
Ross Bennett
}

