% Generated by roxygen2 (4.0.1): do not edit by hand
\name{name.replace}
\alias{name.replace}
\title{utility function to replace awkward named from unlist}
\usage{
name.replace(rnames)
}
\arguments{
\item{rnames}{character vector of names to check for cleanup}
}
\description{
utility function to replace awkward named from unlist
}

