\name{plot.optimize.portfolio.ROI}
\alias{plot.optimize.portfolio.ROI}
\title{plot method for optimize.portfolio.ROI output}
\usage{
  \method{plot}{optimize.portfolio.ROI} (x, ...,
    rp = FALSE, risk.col = "ES", return.col = "mean",
    chart.assets = FALSE, element.color = "darkgray",
    neighbors = NULL, main = "ROI.Portfolios", xlim = NULL,
    ylim = NULL)
}
\arguments{
  \item{x}{object created by
  \code{\link{optimize.portfolio}}}

  \item{...}{any other passthru parameters}

  \item{rp}{TRUE/FALSE to plot feasible portfolios
  generated by \code{\link{random_portfolios}}}

  \item{risk.col}{string matching the objective of a 'risk'
  objective, on horizontal axis}

  \item{return.col}{string matching the objective of a
  'return' objective, on vertical axis}

  \item{chart.assets}{TRUE/FALSE to include risk-return
  scatter of assets}

  \item{element.color}{color for the default plot scatter
  points}

  \item{neighbors}{set of 'neighbor' portfolios to
  overplot}

  \item{main}{an overall title for the plot: see
  \code{\link{title}}}

  \item{xlim}{set the limit on coordinates for the x-axis}

  \item{ylim}{set the limit on coordinates for the y-axis}
}
\description{
  scatter and weights chart for ROI portfolio optimizations
  run with trace=TRUE
}
\details{
  \code{return.col} must be the name of a function used to
  compute the return metric on the random portfolio weights
  \code{risk.col} must be the name of a function used to
  compute the risk metric on the random portfolio weights

  The ROI optimizers do not store the portfolio weights
  like DEoptim or random portfolios random portfolios can
  be generated for the scatter plot.

  \code{return.col} must be the name of a function used to
  compute the return metric on the random portfolio weights
  \code{risk.col} must be the name of a function used to
  compute the risk metric on the random portfolio weights
}
\author{
  Ross Bennett
}
\seealso{
  \code{\link{optimize.portfolio}}
}

