% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/constraint_fn_map.R
\name{pos_limit_fail}
\alias{pos_limit_fail}
\title{function to check for violation of position limits constraints}
\usage{
pos_limit_fail(weights, max_pos, max_pos_long, max_pos_short)
}
\arguments{
\item{weights}{vector of weights to test}

\item{max_pos}{maximum number of assets with non-zero weights}

\item{max_pos_long}{maximum number of assets with long (i.e. buy) positions}

\item{max_pos_short}{maximum number of assets with short (i.e. sell) positions}
}
\value{
TRUE if any position_limit is violated. FALSE if all position limits are satisfied
}
\description{
This is used as a helper function for \code{\link{rp_transform}} to check
for violation of position limit constraints. The position limit constraints
checked are max_pos, max_pos_long, and max_pos_short.
}

