\name{print.constraint}
\alias{print.constraint}
\title{print method for objects of class 'constraint'}
\usage{
  \method{print}{constraint} (x, ...)
}
\arguments{
  \item{x}{object of class constraint}

  \item{\dots}{any other passthru parameters}
}
\description{
  print method for objects of class 'constraint'
}
\author{
  Ross Bennett
}

