\name{print.constraint}
\alias{print.constraint}
\title{print method for constraint objects}
\usage{
  \method{print}{constraint} (x, ...)
}
\arguments{
  \item{x}{object of class \code{constraint}}

  \item{\dots}{any other passthru parameters}
}
\description{
  print method for constraint objects
}
\author{
  Ross Bennett
}

