\name{print.efficient.frontier}
\alias{print.efficient.frontier}
\title{Print an efficient frontier object}
\usage{
  \method{print}{efficient.frontier} (x, ...)
}
\arguments{
  \item{x}{objective of class \code{efficient.frontier}}

  \item{...}{passthrough parameters}
}
\description{
  Print method for efficient frontier objects. Display the
  call to create or extract the efficient frontier object
  and the portfolio from which the efficient frontier was
  created or extracted.
}
\author{
  Ross Bennett
}

