\name{print.optimize.portfolio.DEoptim}
\alias{print.optimize.portfolio.DEoptim}
\title{Printing Output of optimize.portfolio}
\usage{
  \method{print}{optimize.portfolio.DEoptim} (x, ...,
    digits = max(3, getOption("digits") - 3))
}
\arguments{
  \item{x}{an object of class
  \code{optimize.portfolio.DEoptim} resulting from a call
  to \code{\link{optimize.portfolio}}}

  \item{...}{any other passthru parameters}

  \item{digits}{the number of significant digits to use
  when printing.}
}
\description{
  print method for optimize.portfolio.DEoptim
}
\author{
  Ross Bennett
}

