\name{print.optimize.portfolio.GenSA}
\alias{print.optimize.portfolio.GenSA}
\title{Printing Output of optimize.portfolio}
\usage{
  print.optimize.portfolio.GenSA(object,
    digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
  \item{object}{an object of class
  "optimize.portfolio.GenSA" resulting from a call to
  optimize.portfolio}

  \item{digits}{the number of significant digits to use
  when printing}

  \item{...}{any other passthru parameters}
}
\description{
  print method for optimize.portfolio.GenSA
}

