\name{print.optimize.portfolio.pso}
\alias{print.optimize.portfolio.pso}
\title{Printing Output of optimize.portfolio}
\usage{
  print.optimize.portfolio.pso(object,
    digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
  \item{object}{an object of class "optimize.portfolio.pso"
  resulting from a call to optimize.portfolio}

  \item{digits}{the number of significant digits to use
  when printing.}

  \item{...}{any other passthru parameters}
}
\description{
  print method for optimize.portfolio.pso
}

