\name{print.optimize.portfolio.random}
\alias{print.optimize.portfolio.random}
\title{Printing Output of optimize.portfolio}
\usage{
  print.optimize.portfolio.random(x, ...,
    digits = max(3, getOption("digits") - 3))
}
\arguments{
  \item{x}{an object of class
  \code{optimize.portfolio.random} resulting from a call to
  \code{\link{optimize.portfolio}}}

  \item{digits}{the number of significant digits to use
  when printing.}

  \item{...}{any other passthru parameters}
}
\description{
  print method for optimize.portfolio.random
}

