\name{print.optimize.portfolio.rebalancing}
\alias{print.optimize.portfolio.rebalancing}
\title{print method for optimize...}
\usage{print.optimize.portfolio.rebalancing(x, ...)}
\description{print method for optimize.portfolio.rebalancing}
\arguments{\item{x}{object of type optimize.portfolio.rebalancing}
\item{\dots}{any other passthru parameters}}
