% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{print.optimize.portfolio.rebalancing}
\alias{print.optimize.portfolio.rebalancing}
\title{Printing output of optimize.portfolio.rebalancing}
\usage{
\method{print}{optimize.portfolio.rebalancing}(x, ..., digits = 4)
}
\arguments{
\item{x}{an object used to select a method}

\item{digits}{the number of significant digits to use when printing.}

\item{\dots}{any other passthru parameters}
}
\description{
print method for \code{optimize.portfolio.rebalancing} objects
}
\author{
Ross Bennett
}
\seealso{
\code{\link{optimize.portfolio.rebalancing}}
}

