\name{print.portfolio}
\alias{print.portfolio}
\title{Printing Portfolio Specification Objects}
\usage{
  print.portfolio(x, ...)
}
\arguments{
  \item{x}{object of class \code{portfolio}}
}
\description{
  Print method for objects of class \code{portfolio}
  created with \code{\link{portfolio.spec}}
}
\author{
  Ross Bennett
}

