% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.portfolio}
\alias{print.portfolio}
\title{Printing Portfolio Specification Objects}
\usage{
\method{print}{portfolio}(x, ...)
}
\arguments{
\item{x}{an object of class \code{portfolio}}

\item{\dots}{any other passthru parameters}
}
\description{
Print method for objects of class \code{portfolio} created with \code{\link{portfolio.spec}}
}
\author{
Ross Bennett
}
\seealso{
\code{\link{portfolio.spec}}
}

