% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{print.portfolio}
\alias{print.portfolio}
\title{Printing Portfolio Specification Objects}
\usage{
\method{print}{portfolio}(x, ...)
}
\arguments{
\item{x}{an object of class \code{portfolio}}

\item{\dots}{any other passthru parameters}
}
\description{
Print method for objects of class \code{portfolio} created with \code{\link{portfolio.spec}}
}
\author{
Ross Bennett
}
\seealso{
\code{\link{portfolio.spec}}
}

