% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.summary.optimize.portfolio}
\alias{print.summary.optimize.portfolio}
\title{Printing summary output of optimize.portfolio}
\usage{
\method{print}{summary.optimize.portfolio}(x, ...)
}
\arguments{
\item{x}{an object of class \code{summary.optimize.portfolio}.}

\item{...}{any other passthru parameters. Currently not used.}
}
\description{
print method for objects of class \code{summary.optimize.portfolio}
}
\author{
Ross Bennett
}
\seealso{
\code{\link{summary.optimize.portfolio}}
}

