% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.summary.optimize.portfolio.rebalancing}
\alias{print.summary.optimize.portfolio.rebalancing}
\title{Printing summary output of optimize.portfolio.rebalancing}
\usage{
\method{print}{summary.optimize.portfolio.rebalancing}(x, ..., digits = 4)
}
\arguments{
\item{x}{an object of class \code{summary.optimize.portfolio.rebalancing}.}

\item{\dots}{any other passthru parameters}

\item{digits}{number of digits used for printing}
}
\description{
print method for objects of class \code{summary.optimize.portfolio.rebalancing}
}
\author{
Ross Bennett
}
\seealso{
\code{\link{summary.optimize.portfolio.rebalancing}}
}

