\name{random_walk_portfolios}
\alias{random_walk_portfolios}
\title{deprecated random portfolios wrapper until we write a random trades function...}
\usage{random_walk_portfolios(...)}
\description{deprecated random portfolios wrapper until we write a random trades function}
\author{bpeterson}
\arguments{\item{...}{any other passthru parameters}}
