% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/random_portfolios.R
\name{random_walk_portfolios}
\alias{random_walk_portfolios}
\title{deprecated random portfolios wrapper until we write a random trades function}
\usage{
random_walk_portfolios(...)
}
\arguments{
\item{...}{any other passthru parameters}
}
\description{
deprecated random portfolios wrapper until we write a random trades function
}
\author{
bpeterson
}

