% Generated by roxygen2 (4.0.1): do not edit by hand
\name{random_walk_portfolios}
\alias{random_walk_portfolios}
\title{deprecated random portfolios wrapper until we write a random trades function}
\usage{
random_walk_portfolios(...)
}
\arguments{
\item{...}{any other passthru parameters}
}
\description{
deprecated random portfolios wrapper until we write a random trades function
}
\author{
bpeterson
}

