% Generated by roxygen2 (4.0.1): do not edit by hand
\name{randomize_portfolio_v1}
\alias{randomize_portfolio_v1}
\title{Random portfolio sample method}
\usage{
randomize_portfolio_v1(rpconstraints, max_permutations = 200, rounding = 3)
}
\arguments{
\item{rpconstraints}{an object of type "constraints" specifying the constraints for the optimization, see \code{\link{constraint}}}

\item{max_permutations}{integer: maximum number of iterations to try for a valid portfolio, default 200}

\item{rounding}{integer how many decimals should we round to}
}
\value{
named weights vector
}
\description{
This function generates random permutations of a portfolio seed meeting
leverage and box constraints. The final step is to run \code{\link{fn_map}}
on the random portfolio weights to transform the weights so they satisfy
other constraints such as group or position limit constraints. This is the
'sample' method for random portfolios and is based on an idea by Pat Burns.
}
\author{
Peter Carl, Brian G. Peterson, (based on an idea by Pat Burns)
}

