% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.efficient.frontier}
\alias{summary.efficient.frontier}
\title{Summarize an efficient frontier object}
\usage{
\method{summary}{efficient.frontier}(object, ..., digits = 3)
}
\arguments{
\item{object}{object of class \code{efficient.frontier}}

\item{...}{passthrough parameters}

\item{digits}{number of digits to round to}
}
\description{
Summary method for efficient frontier objects. Display the call to create or
extract the efficient frontier object as well as the weights and risk and
return metrics along the efficient frontier.
}
\author{
Ross Bennett
}

