\name{summary.optimize.portfolio}
\alias{summary.optimize.portfolio}
\title{Summarizing Output of optimize.portfolio}
\usage{
  summary.optimize.portfolio(object, ...)
}
\arguments{
  \item{object}{an object of class "optimize.portfolio.pso"
  resulting from a call to optimize.portfolio}

  \item{...}{any other passthru parameters. Currently not
  used.}
}
\description{
  summary method for class "optimize.portfolio"
}

