% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.optimize.portfolio.rebalancing}
\alias{summary.optimize.portfolio.rebalancing}
\title{summary method for optimize.portfolio.rebalancing}
\usage{
\method{summary}{optimize.portfolio.rebalancing}(object, ...)
}
\arguments{
\item{object}{object of type optimize.portfolio.rebalancing}

\item{\dots}{any other passthru parameters}
}
\description{
summary method for optimize.portfolio.rebalancing
}

