\name{summary.portfolio}
\alias{summary.portfolio}
\title{Summary method for objects of class 'portfolio'}
\usage{
  summary.portfolio(portfolio)
}
\arguments{
  \item{portfolio}{object of class portfolio}
}
\description{
  Summary method for objects of class 'portfolio'
}
\author{
  Ross Bennett
}

