% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.portfolio}
\alias{summary.portfolio}
\title{Summarize Portfolio Specification Objects}
\usage{
\method{summary}{portfolio}(object, ...)
}
\arguments{
\item{object}{an object of class \code{portfolio}}

\item{\dots}{any other passthru parameters}
}
\description{
summary method for class \code{portfolio} created with \code{\link{portfolio.spec}}
}
\author{
Ross Bennett
}
\seealso{
\code{\link{portfolio.spec}}
}

