\name{summary.portfolio}
\alias{summary.portfolio}
\title{Summarize Portfolio Specification Objects}
\usage{
  \method{summary}{portfolio} (object, ...)
}
\arguments{
  \item{object}{object of class portfolio}

  \item{\dots}{any other passthru parameters}
}
\description{
  summary method for class \code{portfolio} created with
  \code{\link{portfolio.spec}}
}
\author{
  Ross Bennett
}

