\name{summary.portfolio}
\alias{summary.portfolio}
\title{Summarize Portfolio Specification Objects}
\usage{
  summary.portfolio(object, ...)
}
\arguments{
  \item{object}{object of class portfolio}
}
\description{
  summary method for class \code{portfolio} created with
  \code{\link{portfolio.spec}}
}
\author{
  Ross Bennett
}

