% Generated by roxygen2 (4.0.1): do not edit by hand
\name{update.constraint}
\alias{update.constraint}
\title{function for updating constrints, not well tested, may be broken}
\usage{
\method{update}{constraint}(object, ...)
}
\arguments{
\item{object}{object of type \code{\link{constraint}} to update}

\item{...}{any other passthru parameters, used to call \code{\link{constraint}}}
}
\description{
can we use the generic update.default function?
}
\author{
bpeterson
}

